requirements.txt
wntr
PyYaml
Jinja2
